debug twitter api request

+++ db/migrate/20131208000818_add_screenname_and_phone_to_users.rb
@@ -1,6 +1,6 @@
 class AddScreennameAndPhoneToUsers < ActiveRecord::Migration
   def change
     add_column :users, :twitter_screenname, :string
-    add_column :users, :phone, :integer
+    add_column :users, :phone, :bigint
   end
 end
